refactor: 도메인 폴더 구조 변경
