+++ src/main/java/vendingmachine/domain/coin/Coins.java
@@ -1,4 +1,4 @@
-package vendingmachine.domain;
+package vendingmachine.domain.coin;
 
 import java.util.Map;
