+++ src/test/java/vendingmachine/domain/MenuTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import vendingmachine.domain.menu.Menu;
 import vendingmachine.exceptions.NotDivisibleByMinPriceCoinException;
 
 class MenuTest {
